// src/types.ts
export interface EVMState {
  transactions: TransactionAndReceipt[];
  logs: ParsedLog[];
  providers: ProvidersList;
  parsed: {
    reads: ContractRead[];
    transactions: any;
    logs: any;
  };
}

export interface ParsedLog {
  name: string;
  signature: string;
  topic: string;
  args: any;
  blockNumber: number;
  blockHash: string;
  transactionIndex: number;
  removed?: boolean;
  address: string;
  data: string;
  topics: Array<string>;
  transactionHash: string;
  logIndex: number;
}

type ConstructedStatus =
  | 'sorted' // Sort the Set JSON object for better functional application of set modifiers
  | 'constructed' // Constructs JSON entity object into JS class instances
  | 'injected' // Injects EVM state (transactions, logs and providers) into constructed set
  | 'fetched' // Fetch stateful conditions data from an EVM RPCProvider URI
  | 'parsed' // Parse transaction/log/state bytecode into qualitative entity JS objects/instances
  | 'filtered' // Apply conditionals to computed Entity/State collections
  | string;

export interface ConstructedSmartContractSet {
  collections: EntityCollection[];
  status: ConstructedStatus;
  rules: SmartContractRule[];
}

export interface ConditionMatch {
  conditionId: string;
  stateId: any; // hash/address/etc...
  isComplete: boolean;
  conditions: Array<boolean>;
}

export enum ConditionStateMatchObject {
  Transaction = 'match.transaction',
  Log = 'match.log',
  Read = 'match.read',
}

export interface ConditionStateMatch {
  id: string;
  matches: Array<{
    object: ConditionStateMatchObject;
    conditionId: string;
    stateId: string;
    isComplete: boolean;
    conditions: Array<boolean>;
  }>;
}

export interface EntityCollection {
  entity: SmartContractEntity;
  conditions: SmartContractCondition[];
  interface?: Interface;
  state: EVMState;
  computed: {
    transactions: ConditionStateMatch[];
    logs: ConditionStateMatch[];
    reads: ConditionStateMatch[];
  };
  progress: any;
  metadata: any;
}